re-implement the Send Egg page
